Added cryptography, ACL's, OOP and access modifiers

+++ SSD-CA1/Program.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.IO;
+using System.Security.Cryptography;
+using System.Security.AccessControl;
 
 namespace SSD_CA1
 {
@@ -9,22 +11,30 @@ namespace SSD_CA1
     {
         static void Main(string[] args)
         {
-            string input = "";
-            try
+            using (AesManaged myAes = new AesManaged())
             {
-                do
+                myAes.Padding = PaddingMode.PKCS7;
+                myAes.KeySize = 128;          // in bits
+                myAes.Key = new byte[128 / 8];  // 16 bytes for 128 bit encryption
+                myAes.IV = new byte[128 / 8];   // AES needs a 16-byte IV
+                string input = "";
+                try
                 {
-                    Console.Clear();
-                    input = CheckInput(input);
-                } while (input != "5");
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Exception: " + e.Message);
-                Console.ReadKey();
+                    do
+                    {
+                        Console.Clear();
+                        input = CheckInput(myAes, input);
+                    } while (input != "5");
+                }
+                catch (Exception e)
+                {
+                    Console.WriteLine("Exception: " + e.Message);
+                    Console.ReadKey();
+                }
+                GC.Collect();
             }
         }
-        private static string CheckInput(string input)
+        private static string CheckInput(AesManaged myAes, string input)
         {
             input = "";
             while (input != "5")
@@ -40,16 +50,16 @@ namespace SSD_CA1
                 switch (input)
                 {
                     case "1":
-                        AddData();
+                        AddData(myAes);
                         break;
                     case "2":
-                        GetData(2);
+                        GetData(myAes, 2);
                         break;
                     case "3":
-                        GetData(3);
+                        GetData(myAes, 3);
                         break;
                     case "4":
-                        GetData(4);
+                        GetData(myAes, 4);
                         break;
                     case "5":
                         break;
@@ -60,43 +70,60 @@ namespace SSD_CA1
             }
             return input;
         }
-        private static void AddData()
+        private static void AddData(AesManaged myAes)
         {
             Console.Clear();
             Console.WriteLine("Enter the data");
-            string input = Console.ReadLine();
-            StreamWriter sw = new StreamWriter("Data.txt", true);
-            sw.WriteLine(input);
-            sw.Close();
+            Text input = new Text(Console.ReadLine());
+            // Encrypt the string to an array of bytes.
+            byte[] encrypted = EncryptStringToBytes_Aes(input.value, myAes.Key, myAes.IV);
+            if (File.Exists("Data.txt"))
+            {
+                StreamWriter sw = new StreamWriter("Data.txt", true);
+                sw.WriteLine(Convert.ToBase64String(encrypted));
+                sw.Close();
+            }
+            else {
+                string fileName = "Data.txt";
+                var stream=File.Create(fileName);
+                stream.Close();
+                // Add the access control entry to the file.
+                AddFileSecurity(fileName, Environment.UserDomainName+"\\"+Environment.UserName, FileSystemRights.ReadData, AccessControlType.Allow);
+                StreamWriter sw = new StreamWriter("Data.txt", true);
+                sw.WriteLine(Convert.ToBase64String(encrypted));
+                sw.Close();
+            }
         }
-        private static void GetData(int option)
+        private static void GetData(AesManaged myAes, int option)
         {
             if (File.Exists("Data.txt"))
             {
                 int i = 1;
                 Console.Clear();
-                List<string> arr = new List<string>();
+                List<Text> arr = new List<Text>();
                 Console.WriteLine("Data:");
                 StreamReader sr = new StreamReader("Data.txt");
                 string input = sr.ReadLine();
                 while (input != null && input != "")
                 {
+                    // Decrypt the bytes to a string.
+                    Text decrypted = new Text(DecryptStringFromBytes_Aes(Convert.FromBase64String(input), myAes.Key, myAes.IV));
+                    Console.WriteLine("{0}. {1}", i, decrypted.value);
                     if (option == 3 || option == 4)
                     {
-                        arr.Add(input);
+                        arr.Add(decrypted);
                     }
-                    Console.WriteLine("{0}. {1}", i, input);
                     i++;
                     input = sr.ReadLine();
                 }
                 sr.Close();
                 if (option == 3)
                 {
-                    UpdateData(arr);
+                    UpdateData(myAes, arr);
                 }
                 else if (option == 4)
                 {
-                    DeleteData(arr);
+                    DeleteData(myAes, arr);
                 }
                 else
                 {
@@ -110,7 +137,7 @@ namespace SSD_CA1
                 Console.ReadKey();
             }
         }
-        private static void DeleteData(List<string> arr)
+        private static void DeleteData(AesManaged myAes, List<Text> arr)
         {
             string input = "";
             int intVal;
@@ -123,7 +150,7 @@ namespace SSD_CA1
                     if (intVal <= arr.Count() && intVal > 0)
                     {
                         arr.RemoveAt(intVal - 1);
-                        OverwriteFile(arr);
+                        OverwriteFile(myAes, arr);
                     }
                     else
                     {
@@ -138,7 +165,7 @@ namespace SSD_CA1
                 }
             }
         }
-        private static void UpdateData(List<string> arr)
+        private static void UpdateData(AesManaged myAes, List<Text> arr)
         {
             string input = "";
             int intVal;
@@ -152,8 +179,8 @@ namespace SSD_CA1
                     {
                         Console.WriteLine("What would you like to replace it with");
                         string newInput = Console.ReadLine();
-                        arr[intVal - 1] = newInput;
-                        OverwriteFile(arr);
+                        arr[intVal - 1] = new Text(newInput);
+                        OverwriteFile(myAes, arr);
                     }
                     else
                     {
@@ -168,7 +195,7 @@ namespace SSD_CA1
                 }
             }
         }
-        private static void OverwriteFile(List<string> arr)
+        private static void OverwriteFile(AesManaged myAes, List<Text> arr)
         {
             File.WriteAllText("Data.txt", "");
             if (arr.Count > 0)
@@ -176,13 +203,119 @@ namespace SSD_CA1
                 StreamWriter sw = new StreamWriter("Data.txt", true);
                 foreach (var item in arr)
                 {
-                    sw.WriteLine(item);
+                    // Encrypt the string to an array of bytes.
+                    byte[] encrypted = EncryptStringToBytes_Aes(item.value, myAes.Key, myAes.IV);
+                    sw.WriteLine(Convert.ToBase64String(encrypted));
                 }
                 sw.Close();
             }
-            else {
+            else
+            {
                 File.Delete("Data.txt");
             }
         }
+        static byte[] EncryptStringToBytes_Aes(string plainText, byte[] Key, byte[] IV)
+        {
+            // Check arguments.
+            if (plainText == null || plainText.Length <= 0)
+                throw new ArgumentNullException("plainText");
+            if (Key == null || Key.Length <= 0)
+                throw new ArgumentNullException("Key");
+            if (IV == null || IV.Length <= 0)
+                throw new ArgumentNullException("IV");
+            byte[] encrypted;
+
+            // Create an AesManaged object
+            // with the specified key and IV.
+            using (AesManaged aesAlg = new AesManaged())
+            {
+                aesAlg.Padding = PaddingMode.PKCS7;
+                aesAlg.KeySize = 128;          // in bits
+                aesAlg.Key = new byte[128 / 8];  // 16 bytes for 128 bit encryption
+                aesAlg.IV = new byte[128 / 8];   // AES needs a 16-byte IV
+                aesAlg.Key = Key;
+                aesAlg.IV = IV;
+
+                // Create an encryptor to perform the stream transform.
+                ICryptoTransform encryptor = aesAlg.CreateEncryptor(aesAlg.Key, aesAlg.IV);
+
+                // Create the streams used for encryption.
+                using (MemoryStream msEncrypt = new MemoryStream())
+                {
+                    using (CryptoStream csEncrypt = new CryptoStream(msEncrypt, encryptor, CryptoStreamMode.Write))
+                    {
+                        using (StreamWriter swEncrypt = new StreamWriter(csEncrypt))
+                        {
+                            //Write all data to the stream.
+                            swEncrypt.Write(plainText);
+                        }
+                        encrypted = msEncrypt.ToArray();
+                    }
+                }
+            }
+            GC.Collect();
+            return encrypted;
+        }
+
+        static string DecryptStringFromBytes_Aes(byte[] cipherText, byte[] Key, byte[] IV)
+        {
+            // Check arguments.
+            if (cipherText == null || cipherText.Length <= 0)
+                throw new ArgumentNullException("cipherText");
+            if (Key == null || Key.Length <= 0)
+                throw new ArgumentNullException("Key");
+            if (IV == null || IV.Length <= 0)
+                throw new ArgumentNullException("IV");
+
+            // Declare the string used to hold
+            // the decrypted text.
+            string plaintext = null;
+
+            // Create an AesManaged object
+            // with the specified key and IV.
+            using (AesManaged aesAlg = new AesManaged())
+            {
+                aesAlg.Padding = PaddingMode.PKCS7;
+                aesAlg.KeySize = 128;          // in bits
+                aesAlg.Key = new byte[128 / 8];  // 16 bytes for 128 bit encryption
+                aesAlg.IV = new byte[128 / 8];   // AES needs a 16-byte IV
+                aesAlg.Key = Key;
+                aesAlg.IV = IV;
+
+                // Create a decryptor to perform the stream transform.
+                ICryptoTransform decryptor = aesAlg.CreateDecryptor(aesAlg.Key, aesAlg.IV);
+
+                // Create the streams used for decryption.
+                using (MemoryStream msDecrypt = new MemoryStream(cipherText))
+                {
+                    using (CryptoStream csDecrypt = new CryptoStream(msDecrypt, decryptor, CryptoStreamMode.Read))
+                    {
+                        using (StreamReader srDecrypt = new StreamReader(csDecrypt))
+                        {
+
+                            // Read the decrypted bytes from the decrypting stream
+                            // and place them in a string.
+                            plaintext = srDecrypt.ReadToEnd();
+                        }
+                    }
+                }
+
+            }
+            GC.Collect();
+            return plaintext;
+        }
+
+        public static void AddFileSecurity(string fileName, string account, FileSystemRights rights, AccessControlType controlType)
+        {
+            // Get a FileSecurity object that represents the
+            // current security settings.
+            FileSecurity fSecurity = File.GetAccessControl(fileName);
+
+            // Add the FileSystemAccessRule to the security settings.
+            fSecurity.AddAccessRule(new FileSystemAccessRule(account, rights, controlType));
+
+            // Set the new access settings.
+            File.SetAccessControl(fileName, fSecurity);
+            }
     }
 }
